Refactor Landing Page: Split into Individual Components
- Created separate components for each section of the landing page:
  - HeroSection
  - StatsSection
  - ReasonsSection
  - BenefitsSection
  - CTASection
  - TrainersSection
  - TestimonialsSection
  - FAQSection
  - FoundersSection
- Updated the main LandingPage component to utilize the new structure for improved maintainability and readability.
