+++ src/components/landing/CTASection.jsx
@@ -1,0 +1,18 @@
+export function CTASection() {
+  return (
+    <section className="pt-32 pb-28 bg-sky-950 text-white md:rounded-3xl">
+      <div className="container mx-auto px-4 text-center">
+        <p className="text-sm mb-3">Start Your Journey</p>
+        <h2 className="text-3xl font-bold mb-8">
+          Ready for a Change? Begin Your Wellness Journey!
+        </h2>
+        <button className="bg-gradient-to-r from-blue-500 to-teal-500 text-white px-8 py-3 rounded-full font-medium hover:opacity-90 transition-opacity">
+          Register Now for FREE
+        </button>
+        <p className="mt-4 text-sm text-teal-500">
+          87 Lakhs + already attended
+        </p>
+      </div>
+    </section>
+  );
+}
